+++ database/migrations/2025_02_04_093627_temp_pengiriman_barang_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('id_barang')->nullable();
             $table->string('id_supplier')->nullable();
             $table->integer('qty')->nullable();
-            $table->decimal('harga', 15, 2)->nullable();
-            $table->decimal('total_harga', 15, 2)->nullable();
+            $table->double('harga')->nullable();
+            $table->double('total_harga')->nullable();
             $table->timestamps();
         });
     }
